refactor: extract shared spec-dir loader; split os_apps into submodules
Loader extraction (removes ~830 lines of drifting duplication):
- NEW temper_spec::loader — the single implementation of app/spec directory
  discovery (IOA scan with root/specs precedence, CSDL fallback
  root -> specs/ -> csdl/, sorted Cedar policies, cross-invariants from both
  legacy locations). 6 unit tests; no new dependencies.
- temper-cli serve loader and temper-platform os_apps discovery now consume
  it. Intentional behavior unifications:
  - entity type comes from the parsed automaton name everywhere (the CLI
    derived it from the FILENAME, which latently registered
    test-fixtures/specs/gmail_oauth.ioa.toml as "GmailOauth" instead of the
    "GmailAuth" its automaton declares and temper-server tests expect)
  - duplicate entity-type collisions now error loudly (a HashMap collect
    used to silently drop one)
  - legacy reactions.toml is now rejected by the CLI with the platform's
    migration error (no fixtures use it anywhere in the repo)
  - unparseable specs / unreadable CSDL+Cedar files now surface errors at
    bundle load or deploy instead of silently hiding the app
- Known follow-up: temper-cli verify/codegen and temper-server
  observe/specs/load_dir.rs hold three more filename-based loader copies.

os_apps split (was 2,372 lines after the loader rewire trimmed discovery):
  os_apps/mod.rs        191  decls, public types, read API, re-exports
  os_apps/discovery.rs  498  platform-specific dir scanning (wasm/agents/
                             skills/ADRs/seeds/manifest/guide)
  os_apps/bundle.rs     112  load_app_bundle + dependencies
  os_apps/install.rs    572  install pipeline
  os_apps/bootstrap/    app(203) skills(282) adrs(143) seed(148) fs(318)
Bodies moved verbatim; only visibility widened where a move crosses the new
module boundaries; every external os_apps::X path resolves unchanged.

temper-spec 258, temper-cli 67+70, temper-platform 186 tests green.
DST review: PASS. Code review: PASS.
