+++ crates/temper-platform/src/os_apps/bootstrap/seed.rs
@@ -1,0 +1,148 @@
+//! Seed data bootstrap: creates declared entity instances on first install.
+
+use temper_runtime::tenant::TenantId;
+
+use crate::os_apps::SeedInstance;
+use crate::state::PlatformState;
+
+/// Bootstrap seed data instances into the tenant.
+///
+/// For each seed instance:
+/// 1. Check if the entity type is registered
+/// 2. Create the entity
+/// 3. Dispatch each action in order
+///
+/// Returns descriptions of successfully created instances.
+pub(in crate::os_apps) async fn bootstrap_seed_data(
+    state: &PlatformState,
+    tenant_id: &TenantId,
+    tenant: &str,
+    instances: &[SeedInstance],
+) -> Vec<String> {
+    if instances.is_empty() {
+        return Vec::new();
+    }
+
+    let agent_ctx = temper_server::request_context::AgentContext::for_service("platform-bootstrap");
+    let mut created = Vec::new();
+
+    for instance in instances {
+        // Check if entity type is registered.
+        let type_exists = {
+            let registry = state.registry.read().unwrap(); // ci-ok: infallible lock
+            registry
+                .get_spec(tenant_id, &instance.entity_type)
+                .is_some()
+        };
+        if !type_exists {
+            tracing::warn!(
+                tenant,
+                entity_type = %instance.entity_type,
+                "Skipping seed instance — entity type not registered"
+            );
+            continue;
+        }
+
+        // Determine entity ID.
+        let entity_id = instance.id.clone().unwrap_or_else(|| {
+            // Generate a deterministic ID from type + fields.
+            let hash_input = format!("{}-{}", instance.entity_type, instance.fields);
+            format!(
+                "seed-{}",
+                &format!("{:x}", md5_like_hash(&hash_input))[..12]
+            )
+        });
+
+        // Check if entity already exists (idempotent).
+        if state
+            .server
+            .entity_exists(tenant_id, &instance.entity_type, &entity_id)
+        {
+            tracing::debug!(
+                tenant,
+                entity_type = %instance.entity_type,
+                entity_id = %entity_id,
+                "Seed entity already exists — skipping"
+            );
+            created.push(format!("{}({})", instance.entity_type, entity_id));
+            continue;
+        }
+
+        // Create entity with initial fields.
+        let initial_fields = if instance.fields.is_null() {
+            serde_json::json!({})
+        } else {
+            instance.fields.clone()
+        };
+
+        match state
+            .server
+            .get_or_create_tenant_entity(
+                tenant_id,
+                &instance.entity_type,
+                &entity_id,
+                initial_fields,
+            )
+            .await
+        {
+            Ok(_) => {
+                // Dispatch each action in order.
+                for action in &instance.actions {
+                    let params = if action.params.is_null() {
+                        serde_json::json!({})
+                    } else {
+                        action.params.clone()
+                    };
+                    if let Err(e) = state
+                        .server
+                        .dispatch(temper_server::state::DispatchCommand {
+                            tenant: tenant_id,
+                            entity_type: &instance.entity_type,
+                            entity_id: &entity_id,
+                            action: &action.name,
+                            params,
+                            agent_ctx: &agent_ctx,
+                            await_integration: false,
+                            await_reactions: true,
+                        })
+                        .await
+                    {
+                        tracing::warn!(
+                            tenant,
+                            entity_type = %instance.entity_type,
+                            entity_id = %entity_id,
+                            action = %action.name,
+                            error = %e,
+                            "Failed to dispatch seed action"
+                        );
+                    }
+                }
+                tracing::info!(
+                    tenant,
+                    entity_type = %instance.entity_type,
+                    entity_id = %entity_id,
+                    "Seed entity created"
+                );
+                created.push(format!("{}({})", instance.entity_type, entity_id));
+            }
+            Err(e) => {
+                tracing::warn!(
+                    tenant,
+                    entity_type = %instance.entity_type,
+                    entity_id = %entity_id,
+                    error = %e,
+                    "Failed to create seed entity"
+                );
+            }
+        }
+    }
+    created
+}
+
+/// Simple hash for generating deterministic seed entity IDs.
+fn md5_like_hash(input: &str) -> u64 {
+    use std::hash::{Hash, Hasher};
+    let mut hasher = std::collections::hash_map::DefaultHasher::new();
+    input.hash(&mut hasher);
+    hasher.finish()
+}
